requirements.txt
psycopg2-binary==2.9.10
Flask==3.0.0
Werkzeug==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
Flask-Mail==0.9.1
Authlib==1.6.8
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
yfinance==0.2.32
google-generativeai==0.3.0
python-dotenv==1.0.0
tensorflow==2.14.0
statsmodels==0.14.0
scipy==1.11.0
gunicorn==23.0.0

